Basic worker.
Needs to wait for initialization.
