+++ src/App.tsx
@@ -35,7 +35,7 @@ import BeforeUnloadDirtyCheck from "./workbench/BeforeUnloadDirtyCheck";
 import { SelectionProvider } from "./workbench/use-selection";
 import Workbench from "./workbench/Workbench";
 import ToolkitProvider from "./documentation/toolkit-hooks";
-import SearchProvider from "./documentation/search-hooks";
+import SearchProvider from "./documentation/search/search-hooks";
 
 const isMockDeviceMode = () =>
   // We use a cookie set from the e2e tests. Avoids having separate test and live builds.
