+++ src/documentation/search/SearchResultList.tsx
@@ -1,7 +1,6 @@
 import { Divider, Link, Stack, Text, TextProps } from "@chakra-ui/react";
 import { useRouterState } from "../../router-hooks";
-import { Extract } from "./extracts";
-import { Result } from "./search-hooks";
+import { Extract, Result } from "./common";
 
 interface SearchResultListProps {
   title: string;
